refactor: make the selection slice independent of the map slice and cleanup dead code
