+++ src/features/hotkeys/AppHotkeys.jsx
@@ -8,10 +8,10 @@ import { configure as configureHotkeys, GlobalHotKeys } from 'react-hotkeys';
 import { connect } from 'react-redux';
 
 import { removeSelectedFeatures } from '~/features/map-features/actions';
-import { selectAllUAVs } from '~/features/map/selection';
 import { removeSelectedMissionItems } from '~/features/mission/actions';
 import { getPreferredCommunicationChannelIndex } from '~/features/mission/selectors';
 import { togglePreferredChannel } from '~/features/mission/slice';
+import { selectAllUAVs } from '~/features/selection/slice';
 import {
   toggleBroadcast,
   toggleDeveloperMode,
