+++ src/views/map/MapView.jsx
@@ -34,16 +34,17 @@ import {
   showDetailsForFeatureInTooltipOrGivenFeature,
 } from '~/features/map-features/actions';
 import { getSelectedFeatureIds } from '~/features/map-features/selectors';
+import { getSelectedTool } from '~/features/map/tools';
+import { updateMapViewSettings } from '~/features/map/view';
+import { addNewMissionItem } from '~/features/mission/actions';
+import { getGeofencePolygonId } from '~/features/mission/selectors';
+import { getSelection } from '~/features/selection/selectors';
 import {
   addToSelection,
   removeFromSelection,
   setSelection,
   toggleInSelection,
-} from '~/features/map/selection';
-import { getSelectedTool } from '~/features/map/tools';
-import { updateMapViewSettings } from '~/features/map/view';
-import { addNewMissionItem } from '~/features/mission/actions';
-import { getGeofencePolygonId } from '~/features/mission/selectors';
+} from '~/features/selection/slice';
 import NearestItemTooltip from '~/features/session/NearestItemTooltip';
 import { setFeatureIdForTooltip } from '~/features/session/slice';
 import { getFollowMapSelectionInUAVDetailsPanel } from '~/features/uavs/selectors';
@@ -68,7 +69,6 @@ import {
   getMapViewZoom,
 } from '~/selectors/map';
 import { getVisibleLayersInOrder } from '~/selectors/ordered';
-import { getSelection } from '~/selectors/selection';
 import {
   findFeaturesById,
   lonLatFromMapViewCoordinate,
