+++ src/views/map/layers/mission-info.jsx
@@ -41,6 +41,7 @@ import {
   getMissionItemsWithCoordinatesInOrder,
   getSelectedMissionIndicesForTrajectoryDisplay,
 } from '~/features/mission/selectors';
+import { getSelection } from '~/features/selection/selectors';
 import {
   getConvexHullOfShowInWorldCoordinates,
   getOutdoorShowOrientation,
@@ -57,7 +58,6 @@ import {
 } from '~/model/identifiers';
 import { MissionItemType } from '~/model/missions';
 import { getMapOriginRotationAngle } from '~/selectors/map';
-import { getSelection } from '~/selectors/selection';
 import { hasFeature } from '~/utils/configuration';
 import { formatMissionId } from '~/utils/formatting';
 import { mapViewCoordinateFromLonLat } from '~/utils/geography';
